Fix bug where leftover files from previous versions could prevent code from running
Older versions of Verse used a different format for filenames in localStorage.
Prior to this commit, *all* files were fetched from localStorage and executed
when the page first loaded. This was mostly harmless until the recent change that
makes Verse throw an error if you try to re-`define` a function. Any functions
(e.g. init()) left over in those old files now conflict with the definitions
in main.js, and there's no way to get rid of them from the UI.

This commit fixes the issue by ignoring all files except main.js. Once we do
add the ability to have multiple files in a project, those old ghost files will
reappear, but at that point users will be able to clean them up via the UI, so
it's not as bad.