Move verse.js into a module and migrate Bard tests to Jest
It was getting annoying to maintain the verse framework in
a separate repo, so I'm merging it into this one. I don't
foresee it being used as an independent framework so there's
not much point in having it be in its own repo.

TODO: migrate the rest of the Verse tests from the verse repo.