Save to localStorage 1 second after the last edit
but only if there are no syntax errors. This fixes an annoying issue
where you could lose code if you created an infinite loop or the
browser crashed.

Saving only when the code is syntactically valid is useful because
if you accidentally mess up the syntax and can't figure out how to
fix it, you can refresh to get back to the last good state.