Move docs to subdirectory
The entry points to the docs are now the README.md and the landing page,
which have largely the same content. To save readers the trouble of reading
the same content twice, these two entry points do not link to each other;
both link to a docs/index page, which in turn links to the individual
documentation pages.