Typecheck API functions through wait()
In the course of doing this, I discovered that checkArgs didn't throw
if you pass the wrong number of arguments, so I fixed that, which in
turn broke range(). The easiest way to resolve this was to remove
features of range() which I didn't like anyway. range() now does only
one thing.