Remove log() effect and add output()
The purpose of log() was to display "interstitial" messages when
transitioning from one screen to another. The outgoing screen
can't startDisplay the message, because it's about to terminate; the
destination screen doesn't have the information needed to form
the message.

However, log() also had the incidental effect of showing the whole
history of such messages, which I decided was not worth the added
complexity.

output() uses displayText() and waitForChar() to solve the interstitial
message use case. Spacebar is used as the continuation key because
"press any key to continue" might result in users accidentally skipping
past output they actually cared about. On the other hand, most programs
probably will not use spacebar to e.g. select from a list of options,
so the risk that users will accidentally input something if they reflexively
type Enter-Space to submit something and skip the output is minimized.