Add form() method to simulator
This lets you test that a certain form was displayed.

The implementation of the test framework revealed a subtle bug in the way
formIds are handled. Since formId was previously global, a simulated app
running in a test harness could increment the formId used by the real app. This
caused React to rerender the form whenever tests ran, which meant that you
could be typing in the editor, the tests would run, and the redisplaying form
would steal the focus and cause whatever you typed next to end up in the form's
input box.

To fix this, I scoped the formId to the Process instance. In doing so, I had to
switch to randomly generating formIds instead of the previous incrementing
counter. The motivation for this is that per-process autoincrementing ids can
repeat. If, for example, you restart the app when it's displaying formId 1 and
it then displays formId 1 again immediately after restarting, the form inputs
won't be cleared.