Ensure content can be scrolled into view on small screens
Previously, we positioned the centered container based on viewport
measurements, which isn't correct because a very small viewport puts most of
the content outside the scrollable area (which is based on the size of the
parent Backdrop component).

Now, the Backdrop is always at least as big as the content, the content is
centered within the Backdrop, and if the Backdrop is too big for the viewport,
its parent element gets scrollbars.