Replace Frame component with box-shadow
This makes it much simpler to calculate the dimensions of the CenteredContainer.
Now instead of the magic numbers height=654px and width=1038px we have the
slightly less magic 640 and 1024.

This also allows for easy adjustment of the width of the visual "border"
created by the box-shadow. In the future, I'd like to have different skins that
set different border widths and colors, so this may come in handy.